requirements.txt
absl-py==0.14.0
asgiref==3.4.1
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
clang==5.0
click==8.0.1
colorama==0.4.4
fastapi==0.68.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
gunicorn==20.1.0
h11==0.12.0
h5py==3.1.0
idna==3.2
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
oauthlib==3.1.1
opencv-python-headless==4.5.3.56
opt-einsum==3.3.0
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
python-multipart==0.0.5
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
starlette==0.14.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.7
uvicorn==0.15.0
Werkzeug==2.0.1
wrapt==1.12.1
